Reworked Archive package for standard data types and fixed issue with Server package overrun

+++ engine/server/application/TaskManager/src/shared/GameConnection.cpp
@@ -79,7 +79,7 @@ void GameConnection::receive(const Archive::ByteStream & message)
 	{
 		if(! m_pid)
 		{
-			GenericValueTypeMessage<unsigned long> msg(r);
+			GenericValueTypeMessage<uint32_t> msg(r);
 			m_pid = msg.getValue();
 
 #ifndef _WIN32
